Computer Vision/BLG453E_hw2/requirements.txt
Package         Version 
--------------- --------
autopep8        1.4     
betterpath      0.2.2   
Click           7.0     
cycler          0.10.0  
kiwisolver      1.0.1   
l               0.11.0  
matplotlib      3.0.0   
numpy           1.15.2  
Pillow          5.3.0   
pip             18.1    
pycodestyle     2.4.0   
pyparsing       2.2.2   
PyQt5           5.11.3  
PyQt5-sip       4.19.13 
python-dateutil 2.7.3   
scipy           1.1.0   
setuptools      40.4.3  
six             1.11.0  
vcversioner     2.16.0.0
wheel           0.32.1  
zope.interface  4.5.0   
